+++ importers/CHANGELOG.md
@@ -11,20 +11,47 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
-## [3.2.1] - 2022-09-15
+### Added
+- Abstract DatawellVendor.
+- BlockBuster vendor service.
+- ComicsPlus vendor service.
 
 ### Changed
+- Deprecated "datawell" vendor (now "ComicsPlus" vendor).
+- TheMovieDatabaseVendor has been refactored to use the Abstract DatawellVendor.
+- TheMovieDatabaseVendor now supports single cover no hits processing.
+- Match logic for TheMovieDatabaseVendor has been refactored for better matching.
+- VendorServiceSingleIdentifierInterface now has nullable return.
+- OverDriveMagazinesVendorService has been refactored to use the Abstract DatawellVendor.
+- PressReaderVendorService has been refactored to use the Abstract DatawellVendor.
+- Consolidated all datawell search code to a DataWellClient.
+- Ordered services.yaml and changed "datawell.*" from parameters to direct injection.
+- EbookCentralVendorService has been refactored to use the Abstract DatawellVendor.
+- EbookCentralVendorService now supports single cover no hits processing.
+
+## [3.2.1] - 2022-09-15
+
+### Fixed
 - Fixed vendor id NULL in user upload messages.
 
-## [3.2.0] - 2022-08-22
+## [3.2.0] - 2022-09-12
 
 ### Added
-- Support for hasCover services.
 - "single cover" vendor for no hits processing.
 - "OpenLibrary" single cover vendor.
 
 ### Changed
 - Added "single cover" support to Bogportalen vendor.
+
+## [3.1.1] - 2022-09-03
+
+### Fixed
+- Fixed validate images headers index
+
+## [3.1.0] - 2022-08-22
+
+### Added
+- Support for hasCover services.
 
 ## [3.0.1] - 2022-08-16